// pop-client/src/app/shared/openapi/model/systemic-therapy-create.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SystemicTherapyIntentChoices } from './systemic-therapy-intent-choices';
import { CodedConcept } from './coded-concept';
import { Period } from './period';


export interface SystemicTherapyCreate { 
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Indicates the case of the patient who received the systemic therapy
     */
    caseId: string;
    /**
     * Clinically-relevant period during which the therapy was administered to the patient.
     */
    period: Period;
    /**
     * The total number of treatment cycles during the treatment period.
     */
    cycles: number;
    /**
     * Treatment intent of the system therapy
     */
    intent: SystemicTherapyIntentChoices;
    adjunctiveRole?: CodedConcept | null;
    terminationReason?: CodedConcept | null;
    therapyLineId?: string | null;
    /**
     * References to the neoplastic entities that were targeted by the systemic therapy
     */
    targetedEntitiesIds?: Array<string>;
}
export namespace SystemicTherapyCreate {
}