// pop-client/src/app/shared/openapi/model/systemic-therapy-medication-create.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CodedConcept } from './coded-concept';
import { Measure } from './measure';


export interface SystemicTherapyMedicationCreate { 
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Antineoplastic drug/medication administered to the patient
     */
    drug: CodedConcept;
    route?: CodedConcept | null;
    usedOfflabel?: boolean | null;
    withinSoc?: boolean | null;
    dosageMassConcentration?: Measure | null;
    dosageMass?: Measure | null;
    dosageVolume?: Measure | null;
    dosageMassSurface?: Measure | null;
    dosageRateMassConcentration?: Measure | null;
    dosageRateMass?: Measure | null;
    dosageRateVolume?: Measure | null;
    dosageRateMassSurface?: Measure | null;
}